+++ lerobot/configs/robot/arx5.yaml
